backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.23.2
sqlalchemy>=2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.2
pytest-asyncio==0.22.0
httpx==0.28.1
pytest-cov==4.1.0
greenlet>=3.1.1
